+++ backend/integration_tests/src/client/community.rs
@@ -17,18 +17,21 @@ generate_update_call!(add_reaction);
 generate_update_call!(block_user);
 generate_update_call!(change_role);
 generate_update_call!(create_channel);
+generate_update_call!(create_user_group);
 generate_update_call!(delete_messages);
 generate_update_call!(edit_message);
 generate_update_call!(enable_invite_code);
 generate_update_call!(import_group);
 generate_update_call!(leave_channel);
 generate_update_call!(remove_member);
+generate_update_call!(remove_member_from_channel);
 generate_update_call!(remove_reaction);
 generate_update_call!(send_message);
 generate_update_call!(unblock_user);
 generate_update_call!(undelete_messages);
 generate_update_call!(update_channel);
 generate_update_call!(update_community);
+generate_update_call!(update_user_group);
 
 pub mod happy_path {
     use crate::rng::random_message_id;